Set end of line encoding to Unix (LF)
Sets the end of line encoding in .github\workflows\docker.yml to Unix (LF), to conform with options set in .editorconfig file. This used to be set correctly before, but got changed in one commit because of default TortoiseGit setting to automatically convert EOL between systems. This commit restores the previous, correct encoding.